categories/privacy/metadata-scrubber/requirements.txt
Pillow>=10.0.0
pikepdf>=8.0.0
